Fix compilation on Windows ARM64. THIS PATCH IS BREAKING GREENLET. DO NOT USE!
The purpose of this patch is to make the C extension compilable so that
all the scaffolding for greenlet goes into the site-packages egg.

The .pyd extension is not working and needs to be replaced with a
working one build using https://github.com/mstorsjo/llvm-mingw.

How to set the env to build a working .pyd:

  * Install mingw64 or mingw32
  * Download the llvm-mingw binaries and add them to path
    https://github.com/mstorsjo/llvm-mingw/releases
  * Download the greenlet sources from:
    https://github.com/ader1990/greenlet/tree/llvm-mingw-arm64

How to build the .pyd using llvm-mingw:

  * cd greenlet/src/greenlet
  * Download a Python 3.9 build for ARM64 and unzip here
    https://github.com/ader1990/cpython-builder/actions/runs/377798403
  * aarch64-w64-mingw32-clang -c greenlet.c -o greenlet.o -I./cpython/include
  * aarch64-w64-mingw32-clang -shared -o _greenlet.cp39-win_arm64.pyd greenlet.o -L./cpython/libs -l:python3.lib -l:python39.lib

Now copy _greenlet.cp39-win_arm64.pyd to the place where the greenlet egg was installed.

Example path:

  * cpython\lib\site-packages\greenlet-1.0a2.dev0-py3.9-win-arm64.egg\greenlet\_greenlet.cp39-win_arm64.pyd